Prep for adding projects
