+++ src/components/colored-container/colored-container.module.css
@@ -1,0 +1,3 @@
+.colored-container {
+  transition: background-color 10s;
+}
