+++ src/components/project/project.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import projectStyles from './project.module.css';
+
+const Project = () => <div className={projectStyles.container}>Hello</div>;
+
+export default Project;
