+++ src/components/project/project.module.css
@@ -1,0 +1,6 @@
+.container {
+  background: white;
+  height: 20px;
+  width: 100%;
+  margin-bottom: 30px;
+}
